+++ src/build32.bat
@@ -5,4 +5,4 @@ goto buildlisp
 cl /Zi gen.c
 gen mem.s
 :buildlisp
-cl /Zi flow.c lisp.c cell.c mem32.c math.c /Fe lisp32.exe
+cl /Zi gc.c flow.c lisp.c cell.c mem32.c math.c /Fe lisp32.exe
